requirements.txt
openai>=1.0.0
anthropic>=0.15.0
google-generativeai>=0.3.0
sqlparse>=0.4.3
networkx>=3.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
pymongo>=4.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0
typer>=0.9.0
python-dotenv>=1.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
